Made MenuBuilder items also hideable in the event an entire section needed to be hidden
